requirements.txt
python-dotenv==1.0.1
openai==1.34.0
pydantic==2.7.4
fastapi==0.111.1
retell-sdk==4.0.0
uvicorn==0.30.1